requirements.txt
requests~=2.32.3
PyYAML~=6.0.2
schedule~=1.2.2

typer~=0.15.4